requirements.txt
Django
django-crispy-forms
django-user-agents
Pillow
django-colorfield